% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eek.R
\docType{class}
\name{eek-class}
\alias{eek}
\alias{eek-class}
\title{A Reference Class to represent a bank account.}
\description{
A Reference Class to represent a bank account.
}
\section{Fields}{

\describe{
\item{\code{dat}}{The ECG signal data.}

\item{\code{window}}{Default ECG window when not provided.}

\item{\code{P,Q,R,S,T}}{Peak locations and detection quality.}

\item{\code{bounds}}{Not yet in use.}

\item{\code{startzones}}{Not yet in use.}

\item{\code{endzones}}{Not yet in use.}
}}
\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}

\item{\code{filter(n = 1, lo = 1/40, hi = 1/2)}}{Filter ECG signal using a bandpass filter.}

\item{\code{getPT(minheight = 0, maxrate = 300)}}{Locate all P and T peaks and assign quality score.}

\item{\code{getR(minheight = 0.5, maxrate = 300)}}{Locate all R peaks and assign quality score.}

\item{\code{initialize(file, channel = 1)}}{Initialize eek object using a single ECG channel.}

\item{\code{qplot(view)}}{Visualize an ECG window including peak locations.}

\item{\code{smooth(l = 65, sd = 0.25)}}{Smooth ECG signal by convolving with Gaussian window.}
}}

